+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Fix raycasting scene: Allow setting of number of threads that are used for building a raycasting scene
 * Fix Python bindings for CUDA device synchronization, voxel grid saving (PR #5425)
 * Support msgpack versions without cmake
+* Introduce new optional dependencies support, allowing you to do `pip install open3d[standard]`
 
 ## 0.13